net:mana: Query device and decides msi sharing for EQs
When querying the device, adjust the max number of queues to allow
dedicated MSIx vectors for each vPort. The number of queues per vPort
is clamped down to no less than 16. MSIx sharing among vPort is disabled
by default and it is turned on only when it's not possible.

merge_query